+++ src/main/java/com/avrix/Launcher.java
@@ -1,10 +1,14 @@
 package com.avrix;
 
 import com.avrix.agent.AgentLoader;
+import com.avrix.logs.LineReadingOutputStream;
 import com.avrix.plugin.PluginManager;
 import com.avrix.utils.PatchUtils;
+import org.tinylog.Logger;
 import zombie.gameStates.MainScreenState;
 import zombie.network.GameServer;
+
+import java.io.PrintStream;
 
 /**
  * The Launcher class serves as the entry point for the application.
@@ -17,6 +21,9 @@ public class Launcher {
      * @throws Exception if a critical error occurs that prevents startup
      */
     public static void main(String[] args) throws Exception {
+        System.setOut(new PrintStream(new LineReadingOutputStream(Logger::info), true));
+        System.setErr(new PrintStream(new LineReadingOutputStream(Logger::error), true));
+
         AgentLoader.loadAgent();
 
         PatchUtils.applyDefaultPatches();